+++ scripts/db_migrations/000002_create_tokens_table.up.sql
@@ -4,6 +4,6 @@ CREATE TABLE IF NOT EXISTS tokens
     created_at TIMESTAMPTZ NOT NULL,
     project VARCHAR(80) NOT NULL,
     CONSTRAINT tokens_pkey PRIMARY KEY (token_id),
-    FOREIGN KEY (project) REFERENCES projects(project)
+    FOREIGN KEY (project) REFERENCES projects(project) on delete cascade on update cascade
 );
 GRANT ALL PRIVILEGES ON tokens TO cello;