+++ service/internal/db/db.go
@@ -4,6 +4,7 @@ package db
 
 import (
 	"context"
+	"time"
 
 	"github.com/upper/db/v4"
 	"github.com/upper/db/v4/adapter/postgresql"
@@ -24,6 +25,7 @@ type TokenEntry struct {
 type Client interface {
 	CreateProjectEntry(ctx context.Context, pe ProjectEntry) error
 	DeleteProjectEntry(ctx context.Context, project string) error
+	CreateTokenEntry(ctx context.Context, project string, secretAccessor string) (TokenEntry, error)
 	ListTokenEntries(ctx context.Context, project string) ([]TokenEntry, error)
 	ReadProjectEntry(ctx context.Context, project string) (ProjectEntry, error)
 }
@@ -113,4 +115,30 @@ func (d SQLClient) DeleteProjectEntry(ctx context.Context, project string) error
 	defer sess.Close()
 
 	return sess.WithContext(ctx).Collection(ProjectEntryDB).Find("project", project).Delete()
+}
+
+func (d SQLClient) CreateTokenEntry(ctx context.Context, project string, secretAccessor string) (TokenEntry, error) {
+	res := TokenEntry{}
+
+	sess, err := d.createSession()
+	if err != nil {
+		return res, err
+	}
+	defer sess.Close()
+
+	err = sess.WithContext(ctx).Tx(func(sess db.Session) error {
+		res = TokenEntry{
+			CreatedAt: time.Now().UTC().Format(time.RFC3339),
+			ProjectID: project,
+			TokenID:   secretAccessor,
+		}
+
+		if _, err = sess.Collection(TokenEntryDB).Insert(res); err != nil {
+			return err
+		}
+
+		return nil
+	})
+
+	return res, err
 }