+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Added schema updates to create tokens table
 * Added golang-migrate to manage schema
 * List project tokens
+* Create project token
 
 ## [0.13.3] - 2022-06-16
 ### Changed